+++ src/settings/ShellCommandConfiguration.ts
@@ -6,6 +6,7 @@ export interface ShellCommandConfiguration {
     shell_command: string;
     alias: string;
     confirm_execution: boolean;
+    ignore_error_codes: number[];
 }
 
 export function newShellCommandConfiguration(shell_command: string = ""): ShellCommandConfiguration {
@@ -13,5 +14,6 @@ export function newShellCommandConfiguration(shell_command: string = ""): ShellC
         shell_command: shell_command,
         alias: "",
         confirm_execution: false,
+        ignore_error_codes: [],
     }
 }